statistics aside is added
