+++ bash.exe.stackdump
@@ -1,0 +1,9 @@
+Stack trace:
+Frame        Function    Args
+000FFFFA368  0018006021E (00180252DED, 001802340A6, 000FFFFA368, 000FFFF9260)
+000FFFFA368  00180048859 (00000000000, 00000000000, 00000000000, 00000000000)
+000FFFFA368  00180048892 (00180252EA9, 000FFFFA218, 000FFFFA368, 00000000000)
+000FFFFA368  001800AF0D8 (00000000000, 00000000000, 00000000000, 00000000000)
+000FFFFA368  001800AF25D (000FFFFA380, 00000000000, 00000000000, 00000000000)
+000FFFFA5F0  001800B0673 (000FFFFA380, 00000000000, 00000000000, 00000000000)
+End of stack trace
