+++ src/components/Statistics/Statistic.module.css
@@ -1,5 +1,16 @@
+.wrapper {
+  display: flex;
+}
+
+.aside {
+  width: 52px;
+  min-height: calc(100vh - 52px);
+  box-shadow: 2px 3.464px 8px 0px rgba(29, 29, 27, 0.1);
+}
+
 .outer {
-  max-width: 375px;
+  width: 80%;
+  /* max-width: 375px; */
   margin: 0 auto;
 }
 
@@ -35,13 +46,13 @@
 }
 
 @media only screen and (min-width: 470px) {
-  .outer {
+  /* .outer {
     max-width: 768px;
   }
 
   .inner {
     max-width: 470px;
-  }
+  } */
 
   .chartWraper {
     width: 360px;
@@ -50,11 +61,13 @@
 
 @media only screen and (min-width: 942px) {
   .outer {
-    max-width: 1200px;
+    /* max-width: 1200px; */
+    width: 80%;
   }
 
   .inner {
     max-width: 942px;
+    width: 100%;
   }
 
   .chartWraper {
@@ -68,6 +81,7 @@
   }
 
   .componentsWraper {
+    width: 100%;
     display: flex;
     flex-direction: row;
     justify-content: space-between;
